updated azure credentials in config
