--- .core_cache/temp_model_files/model/MLmodel
@@ -1,17 +1,0 @@
-flavors:
-  catboost:
-    catboost_version: 1.0.5
-    code: null
-    data: model.cb
-    model_type: CatBoostClassifier
-    save_format: cbm
-  python_function:
-    data: model.cb
-    env:
-      conda: conda.yaml
-      virtualenv: python_env.yaml
-    loader_module: mlflow.catboost
-    python_version: 3.9.0
-mlflow_version: 2.3.2
-model_uuid: 76957b55dd7145a0a723d2c638db79fd
-utc_time_created: '2023-08-22 15:37:13.314378'
