--- .core_cache/temp_model_files/model/conda.yaml
@@ -1,13 +1,0 @@
-channels:
-- conda-forge
-dependencies:
-- python=3.9.0
-- pip<=23.0.1
-- pip:
-  - mlflow==2.3
-  - catboost==1.0.5
-  - ipython==8.13.2
-  - numpy==1.21.0
-  - pandas==1.5.3
-  - scipy==1.10.1
-name: mlflow-env
