--- .core_cache/temp_model_files/model/python_env.yaml
@@ -1,7 +1,0 @@
-python: 3.9.0
-build_dependencies:
-- pip==23.0.1
-- setuptools==67.8.0
-- wheel==0.38.4
-dependencies:
-- -r requirements.txt
