--- .core_cache/temp_model_files/model/requirements.txt
@@ -1,6 +1,0 @@
-mlflow==2.3
-catboost==1.0.5
-ipython==8.13.2
-numpy==1.21.0
-pandas==1.5.3
-scipy==1.10.1
